move vectorDb to services and give more descriptive name.
