+++ src/services/vectorDb/VectorDB.ts
@@ -1,8 +1,8 @@
 import Database from 'better-sqlite3';
 import vectorlite from 'vectorlite';
 import { OpenAI } from 'openai';
-import { config } from '../config/index.js';
-import { createLogger } from '../utils/logger.js';
+import { config } from '../../config/index.js';
+import { createLogger } from '../../utils/logger.js';
 import { join } from 'path';
 import { existsSync, mkdirSync } from 'fs';
